Develop

+++ src/AppBundle/Controller/CarController.php
@@ -12,19 +12,33 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 
 class CarController extends Controller
 {
+    /**
+     * Show form for edit car
+     * 
+     * @param idCar $idCar
+     */
     public function editAction($idCar)
     {
-        $security_context = $this->get('security.context');
+        $em = $this->getDoctrine()->getEntityManager();
+        $security_context = $this->get('security.token_storage');
         $security_token = $security_context->getToken();
         $user = $security_token->getUser();
         $companyId = $user->getCompanys()->getId();
-        $car = $this->getDoctrine()->getEntityManager()->getRepository("AppBundle:car")->findOneById($idCar);        
+        $companyService = $this->container->get("company_service");
+        $car = $em->getRepository("AppBundle:car")->findOneById($idCar);        
+        if($car == null)
+            return $this->redirectToRoute('show_car');
+        if($companyService->existCar($companyId,$car->getId()))
+            return $this->redirectToRoute('show_car');
         return $this->render('car/content-panel-createCar.html.twig', array("car"=>$car,"user_type" => "company", "companyId" => $companyId,"carId" => $car->getid()));
-        //return $this->redirect('login');
     }
 
+    /**
+     * Edit car
+     * 
+     * @param Request $request
+     */
     public function editAjaxAction(Request $request){
-
         $em = $this->getDoctrine()->getEntityManager();
         $plate = $request->get('plate');
         $trademark = $request->get('trademark');
@@ -32,18 +46,20 @@ class CarController extends Controller
         $version = $request->get('version');
         $state = $request->get('state');
         $idCar = $request->get('idCar');
-
         $response = array(
             "status" => false,
             "message" => "ERROR"
         );
-
-        if (empty($plate) or empty($trademark) or empty($model) or empty($version) or empty($state) or empty($idCar)){
+        if(!preg_match("/^\d{4}[A-Z]{3}/", $plate)){
             $response = array(
                 "status" => false,
-                "message" => "rellene los campos"
+                "message" => "Formato de matrícula incorrecto"
             );
-            
+        }else if (empty($plate) or empty($trademark) or empty($model) or empty($version) or empty($state) or empty($idCar)){
+            $response = array(
+                "status" => false,
+                "message" => "Rellene los campos"
+            );
         }else{
             try{
                 $car = $em->getRepository("AppBundle:Car")->findOneById($idCar);
@@ -69,12 +85,30 @@ class CarController extends Controller
         return new JsonResponse($response);
     }
 
-    public function deleteAction(Request $request,$idCar){
+    /**
+     * Set state for driver
+     * 
+     * @param Request $request
+     */
+    public function deleteAction(Request $request){
+        $idCar = $request->get('idCar');
         $em = $this->getDoctrine()->getEntityManager();
-        $car = $em->getRepository("AppBundle:car")->findOneById($idCar);
+        $security_context = $this->get('security.token_storage');
+        $security_token = $security_context->getToken();
+        $user = $security_token->getUser();
+        $idCompany = $user->getCompanys()->getId();
+        $companyService = $this->container->get("company_service");
+        $car = $em->getRepository("AppBundle:car")->findOneById($idCar);        
+        if($car == null)
+            return $this->redirectToRoute('show_car');
+        if($companyService->existCar($idCompany,$car->getId()))
+            return $this->redirectToRoute('show_car');
         $em->remove($car);
         $em->flush();
-        return $this->redirect($this->generateUrl('show_car'));
+        return new JsonResponse(array(
+            "status" => true,
+            "message" => "El coche ha sido eliminado"
+        ));
     }
 
 }
